better if record method
