+++ FatCat.Fakes/DynamicRecordCreator.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.CompilerServices;
 
 namespace FatCat.Fakes;
 
@@ -32,28 +31,6 @@ internal static class DynamicRecordCreator
 
     public static bool IsRecordType(this Type type)
     {
-        if (!type.IsClass && !type.IsValueType)
-        {
-            return false; // Must be a class or struct
-        }
-
-        // Check if ToString is overridden
-        var toStringMethod = type.GetMethod("ToString", Type.EmptyTypes);
-        var hasToStringOverride = toStringMethod?.DeclaringType == type;
-
-        // Check if Equals is overridden
-        var equalsMethod = type.GetMethod("Equals", new[] { typeof(object) });
-        var hasEqualsOverride = equalsMethod?.DeclaringType == type;
-
-        // Check if any property has an 'init' accessor (unique to records)
-        var hasInitOnlyProperties = type.GetProperties()
-            .Any(
-                prop =>
-                    prop.SetMethod
-                        ?.ReturnParameter?.GetRequiredCustomModifiers()
-                        .Contains(typeof(IsExternalInit)) == true
-            );
-
-        return (hasToStringOverride && hasEqualsOverride) || hasInitOnlyProperties;
+        return type.GetMethods().Any(m => m.Name == "<Clone>$");
     }
 }
