+++ FatCat.Fakes/FatCat.Fakes.csproj
@@ -12,7 +12,7 @@
         <RepositoryUrl>https://github.com/DavidBasarab/FatCat.Fakes</RepositoryUrl>
         <NoWarn>CS8002,CS1700</NoWarn>
         <SignAssembly>true</SignAssembly>
-        <TargetFrameworks>net8.0;net7.0</TargetFrameworks>
+        <TargetFramework>net9.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
