+++ OneOff/BaseItem.cs
@@ -1,0 +1,8 @@
+namespace OneOff;
+
+public abstract class BaseItem
+{
+    // public string Name { get; set; }
+    //
+    // public int Number { get; set; }
+}
