+++ OneOff/GenericClass.cs
@@ -1,0 +1,7 @@
+namespace OneOff;
+
+public class GenericClass<T> : ModuleItem
+    where T : BaseItem
+{
+    public T Value { get; set; }
+}
