+++ OneOff/IdObjectLevel.cs
@@ -1,0 +1,8 @@
+using MongoDB.Bson;
+
+namespace OneOff;
+
+public class IdObjectLevel : BottomLevel
+{
+    public ObjectId Id { get; set; }
+}
