+++ OneOff/ItemOne.cs
@@ -1,0 +1,6 @@
+namespace OneOff;
+
+public class ItemOne : BaseItem
+{
+    // public string ItemOneProperty { get; set; }
+}
