+++ OneOff/ItemTwo.cs
@@ -1,0 +1,6 @@
+namespace OneOff;
+
+public class ItemTwo : BaseItem
+{
+    // public int ItemTwoProperty { get; set; }
+}
