+++ OneOff/Models/SpaceCraftStatus.cs
@@ -1,0 +1,11 @@
+namespace OneOff.Models;
+
+public enum SpaceCraftStatus
+{
+    UnderConstruction,
+    Cancelled,
+    Destroyed,
+    Complete,
+    Mothballed,
+    Lost,
+}
