+++ OneOff/ModuleItem.cs
@@ -1,0 +1,8 @@
+namespace OneOff;
+
+public abstract class ModuleItem
+{
+    public string Room { get; set; }
+
+    public int Times { get; set; }
+}
