+++ OneOff/TestingRecord.cs
@@ -1,0 +1,3 @@
+namespace OneOff;
+
+public record TestingRecord(int Number, string Name);
